Support retrieving custom Slack metadata from token
Adds a CustomExchange config pointer that allows a provider to replace
the default oauth2.Exchange routine with a custom implementation that
returns a metadata string in addition to the normal Token. The metadata
string is returned back to the caller in the response alongside the
sealed parameter.

The slack profile uses this in conjunction with the slack-go package to
return a copy of the additional metadata fields returned by the
oauth.v2.access API (stripped off all tokens).

Also upgrades the Slack profile to use the /v2 Slack OAuth endpoints.